+++ CHANGELOG.md
@@ -1,9 +1,24 @@
 # Change Log
 
-All notable changes to the "pythonclass" extension pack will be documented in this file.
+All notable changes to the "vscode-python-class" extension will be documented
+in this file.[^1]
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+0.1.0 -- 2022-03-02
+-------------------
 
-- Initial release
+### Added
+
+- `Python Class: Setup Poetry` command
+
+
+0.0.1 -- 2022-02-01
+-------------------
+
+### Added
+
+- Initial release--basic extension pack
+
+
+[^1]: Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how
+to structure this file.